Fixes missing gene-drugs bug
The gene-drugs relations without associated cancer types were not shown in the result table. The cause of this bug was that the filter that removes the unselected gene types from the result list was removing also the gene-drugs without any cancer type associated. Now, this kind of gene-drugs are always included when all the cancer types are selected.