Adds a reverse proxy to bypass cors restriction to backend
If the backend has the CORS filter disabled, that is, it disallows
requests from other domains, the grunt server enables a proxy, so
all requests that can't be served by the front-end, are proxied to
the backend. We can then use the front-end url as the backend
url. Requests to /pandrugsdb-backend, which are not recongnized by
the frontend, are proxied to localhost:8080/pandrugsdb-backend