Updates the backend URLs to the new ones after reorganization
This commit updates the backend URLs to the new URLs after the reorganization
done in the backend. In addition, as the backend has changed the attribute
naming format to camel case, now the javascript attributes can be directly
referenced with the dot operator and the square braces are no longer required.

As an additional change, a BACKEND constant has been added to the app
configuration. This constant contains three values (SERVER, APP and API),
with the base host and URLs to be used when querying the backend.