Modifies the drug query panel
The drug query panel has been changed to allow only single drug queries and to
show more information about the drug name. It also adapts the request to the
backend to retrieve the complete drug names (standard, show and sources) and to
allow queries by any of the drug names.