Adds help
Adss the help infrastructure, based on markdown in order to make
easy to integrate the work made by the members in charge of
documentation.

The help is not finished yet, but the app is not finished itself,
so this commit mainly shows how the help will be integrated in
the app.