Modifies the drug query result view to present a single gene drug group
The drug queries now are done by a single drug. This commit updates the
presentation of the results to show the gene drug group returned expanded.

In order to improve the results presentation, the style of the result view has
been modified and now the best candidate groups and gene drugs have different
color.