Adds more E2E tests
In addition, the backend for E2E tests is configurable on Gruntfile.js,
changing configureProxies:serverE2ETest with configureProxies:server in
the 'test' task of Grunt.

For the remote backend to work, we have increased the timeouts in
protractor.conf.js.

Run tests with: grunt test
Remember to run: ./node_modules/protractor/bin/webdriver-manager start