Adapts the gene symbol retrieval to the new backend API
The backend API has been updated to include more information in the gene-drug
queries. One of the main changes is the increase of the information provided
for each gene and, instead of returning only its gene symbol, now the Entrez
ids and/or the pathways related are also returned.

This commit updates the frontend to retrieve the information required from the
new response format used.