Adds support for indirect resistances
Some genes may have an indirect resistance to drugs caused for the
overactivation of another genes. This may cause a gene sensible to a drug to
act as if it have a resistance.

This commit adds this information to the results table, marking the indirect
resistances and informing the user of their origin.