Adjusts column size in results table
The "Drug status" and "Family" columns of the result table, depite of not being
the most important, usually have the biggest width, as they use to have the
biggest content.

This commit reduces the maximum size of these columns by limiting the maximum
size, increasing the space available for other columns (specially, for the drug
column).