Moves jQuery file upload format assignment to fileupload directive
The file input fields used in generank and vcfrank queries use the fileModel
directive to manage the file selection. In addition, in the controllers of the
query components a jQuery format was assigned to the input field.

This commit moves the jQuery format assignment from the component controllers
to the directive.