Fixes gene dropdown selection in gene query panel
The dropdown selection of genes in the gene query panel was not properly
working, as it didn't update the model value.

This commit fixes this problem by forcing an update event when the text area
is focused. This fixes the problem, as after selecting a gene in the dropbox
the text area gains focus.