Optimizes query execution
Currently, each query was requested twice to the server, as the "Query" button
performs a request and the results table repeats the same request. In addition,
each time the table is reordered or a page is changed, a request is also done.

This commit changes this behaviour by caching the query result and making the
table use the cached result instead of doing request to the server.