Replaces results table in drug queries for a custom table
The drug queries only have one single gene-drug group, as each group is
associated to a single drug and users can only query for a single drug.

In this basis, the results table has been replaced for a custom table for
drug queries, that only shows gene-drug information. The important gene-drug
group information has been moved to a information panel on top of the results
table.

To implement this the results table was converted into a component and the new
drugs result table was also created as a component.