Fixes some problems execution problems when using the distribution version
Several changes where not working when the application was packaged for
distribution.

Component templates could not be found, as Grunt configuration does not include
HTML file that are not in the views directories. Component's HTML files where
moved to views/component/<component-name> directories. In addition, Grunt was
configure to include HTML files in subdirectories of views directory.

The drug query panel was not working due to an missing "bind(this)" in a
function.

A minor syntax problem and a "undefined" check were added to the vcf query
panel component.

Although it is not a packaging problem, a new input parameter was added to the
gene table component to avoid showing variant information in "non-vcf" queries
done after a VCF computation was selected.