Fixes cancer types not being used in queries
The cancer types where not currently used in queries. This is a regression bug
introduced during refactoring.

Although the cancer type filtering was previously done in the frontend, now it
is done by the backend, so the selected cancer types are now sent to the
backend as query parameters.

In addition, the advanced options panel has been improved, and the cancer types
subpanel is disabled when no cancer drug status is selected.