Fixes dropdown drug list not being updated on drug selection
When a drug was selected in the dropdown list of the drug query panel the
input field was updated but the dropdown list was not updated with the new
value.

This commit fixes this problem by listening the "drug" value changed (it is
binded to the input field) and updating the dropdown list when it is changed.