Adds capabilities to exploit the VCF mail notifications
When a VCF computation finishes, the backend sends an email to the
user. A new parameter can be sent to the backend in order to include
a link to the frontend for the user to see results (resultsurltemplate).

This patch sends this parameter with the following link:
http://pandrugs-host/login?ref=query?tab=vcfrank

This commit adds new behaviours to work with this link as follows:
- The login view uses the "ref" parameter to redirect the user to that
route when the login succeeds (query?tab=vcfrank). If the user is
already logged, the redirection is immediate.
- After the redirection to query?tab=vcfrank, the query view is shown
and the vcfrank tab is selected.