Shows interaction types on results bubble chart
In order to show the interaction type of each therapy, the shape
of the figures changes among different interaction types. In fact,
the chart series is now the interaction type, whereas the drug
status is used only to set the color of each point (it is not a serie
anymore).

Highcharts and highcharts-ng were also updated, making some minor
changes necessary, in order to make it to work again. Altough this
upgrade was not strictly necessary, it gives some additional features
that could be used in the future.