Allows filtering queries by the interaction type
Currently, the advanced options menu only allowed filtering the results
by the interaction type based on if it is direct/indirect and
target/marker. However, the results table show direct target, biomarker
and pathway member interactions.

This commit modifies the advanced options panel to allow filtering the
results by the same interactions types as the shown in the result table.