Adds a load from file functionality to gene list query
This commit allows the user to load a gene list from a file in
the genes query. The file is loaded locally and its contents are
placed in the textarea. The file is validated by ensuring that
it contains lines with no commas, semicolons nor intermediate spaces.

In addition, it changes http to https for backend server