Fixes not showing biomarker points in bubble chart and other minor changes
The bubble chart was not showing biomarker relations, since a change in
services/queryresult.js renamed the "marker" relation with "biomarker".
Now the bubble chart is updated to also take into account this change.

In addition, the series are reordered (target, biomarker, pathway member),
and the red lines dashed lines highlighting the best candidate area are
colored in gray.