Fixes bower dependency versions
This commit fixes bower dependencies (bower.json) to known-to-work
specific versions. In addition, it removes the "angular-list-group"
dependency, that seems not to be used and it is not available in
github.com. We should pay attention if there is some strange behaviour
in the future that could be related to this dependency removal.