Fixes drug list not shown
The drug search was not displaying a list of drugs when typing text
in the search field. This was due to a missing dependency whose
repository had disappeared from GitHub.

This commit resolves this issue by incorporating the dependency and
changing the GitHub repository of the dependency to one self-hosted
by the SING group.