Fixes regression in commit 8adf10ca for automatic tab selection
When the guest user submits a computation a link is shown in order
to be kept safe by this user for coming back to see the computation.

In addition, the browser is redirected to this link in order to make
easier to copy this url from the browser location bar.

However, since commit 8adf10ca, this link is not valid does not work,
and it does not selects the VCF tab.

This commit fixes this regression.