Adds initial report infrastructure
This commit adds a new component, called <report> which is used
alongside the <gene-results-table>.

This component shows a button that brings a modal dialog with the
report contents. This dialog has a print button that invokes
window.print(). With the report.css and the media query "print", all the
PanDrugs interface is hidden, with the exception of the report contents. So,
it is straightforward to save the report into PDF or print it.