Adds a skeleton report.
Added a template for the layout. All values are static, with the exception
of the number of input gene list, showing how to get values from the backing
controller of the report component.

Minor changes has been also added. Ng-if is used to avoid execution of many
dom rendering which is not shown until the modal report is brought up.