Removes HTML from Markdown pages
This commit removes the HTML tags and converts the elements to
markdown. To do this, it is necessary to modify the images so
that they are seen with the appropriate size since styles cannot
be applied to the Markdown page.

The Help page has not been modified due to the large number of HTML
tags and links, and since it works on the web, it is not modified.